feat!: trim the package root; raw schemas move to fitzroy/schemas (MNT-04a)
index.ts exported ~120 symbols including every raw AFL API Zod schema
and deprecated aliases, making each a semver commitment and letting
upstream wire drift force major releases. The root now carries only the
supported surface; raw upstream schemas live behind the new
fitzroy/schemas subpath (own esbuild bundle + dts), and the deprecated
date-parse and SquadPlayer aliases are deleted.
