+++ test/lib/date-utils.test.ts
@@ -1,16 +1,8 @@
 import { describe, expect, it } from "vitest";
-import {
-  localToUtc,
-  parseAflApiDate,
-  parseAflApiMatchTime,
-  parseAflTablesDate,
-  parseDate,
-  parseFootyWireDate,
-  toAestString,
-} from "../../src/lib/date-utils";
+import { localToUtc, parseDate, toAestString } from "../../src/lib/date-utils";
 import { DstGapError } from "../../src/lib/errors";
 
-describe("parseDate", () => {
+describe("parseDate (canonical)", () => {
   // AFL API — ISO without Z (utcStartTime)
   it("parses AFL API datetime without Z as UTC", () => {
     expect(parseDate("2026-03-05T08:30:00")?.toISOString()).toBe("2026-03-05T08:30:00.000Z");
@@ -57,56 +49,48 @@ describe("parseDate", () => {
   });
 });
 
-describe("parseAflApiDate", () => {
+describe("parseDate — AFL API ISO inputs", () => {
   it.each([
     ["2024-03-14T06:20:00.000Z", "2024-03-14T06:20:00.000Z"],
     ["2024-03-14T06:20:00Z", "2024-03-14T06:20:00.000Z"],
     ["2024-03-14", undefined], // valid but year-only check below
   ])("parses %s", (input, expectedIso) => {
-    const date = parseAflApiDate(input);
+    const date = parseDate(input);
     expect(date).toBeInstanceOf(Date);
     if (expectedIso) expect(date?.toISOString()).toBe(expectedIso);
     else expect(date?.getUTCFullYear()).toBe(2024);
   });
 
   it.each(["not-a-date", ""])("returns null for %j", (input) => {
-    expect(parseAflApiDate(input)).toBeNull();
+    expect(parseDate(input)).toBeNull();
   });
 });
 
-describe("parseAflApiMatchTime", () => {
+describe("parseDate — AFL API match times", () => {
   it("parses UTC string without Z suffix correctly", () => {
     // AFL API returns UTC times without Z — must not be treated as local time
-    expect(parseAflApiMatchTime("2026-03-05T08:30:00")?.toISOString()).toBe(
-      "2026-03-05T08:30:00.000Z",
-    );
+    expect(parseDate("2026-03-05T08:30:00")?.toISOString()).toBe("2026-03-05T08:30:00.000Z");
   });
 
   it("parses UTC string with Z suffix correctly", () => {
-    expect(parseAflApiMatchTime("2026-03-05T08:30:00Z")?.toISOString()).toBe(
-      "2026-03-05T08:30:00.000Z",
-    );
+    expect(parseDate("2026-03-05T08:30:00Z")?.toISOString()).toBe("2026-03-05T08:30:00.000Z");
   });
 
   it("parses UTC string with milliseconds", () => {
-    expect(parseAflApiMatchTime("2026-07-04T09:30:00.000Z")?.toISOString()).toBe(
-      "2026-07-04T09:30:00.000Z",
-    );
+    expect(parseDate("2026-07-04T09:30:00.000Z")?.toISOString()).toBe("2026-07-04T09:30:00.000Z");
   });
 
   it("strips timezone offset if present", () => {
-    expect(parseAflApiMatchTime("2026-03-05T08:30:00+00:00")?.toISOString()).toBe(
-      "2026-03-05T08:30:00.000Z",
-    );
+    expect(parseDate("2026-03-05T08:30:00+00:00")?.toISOString()).toBe("2026-03-05T08:30:00.000Z");
   });
 
   it("returns null for invalid input", () => {
-    expect(parseAflApiMatchTime("not-a-date")).toBeNull();
-    expect(parseAflApiMatchTime("")).toBeNull();
+    expect(parseDate("not-a-date")).toBeNull();
+    expect(parseDate("")).toBeNull();
   });
 });
 
-describe("parseFootyWireDate", () => {
+describe("parseDate — FootyWire inputs", () => {
   it.each([
     ["Sat 16 Mar 2024", "2024-03-16T00:00:00.000Z"],
     ["16 Mar 2024", "2024-03-16T00:00:00.000Z"],
@@ -116,41 +100,41 @@ describe("parseFootyWireDate", () => {
     ["16 March 2024", "2024-03-16T00:00:00.000Z"],
     ["  16 Mar 2024  ", "2024-03-16T00:00:00.000Z"],
   ])("parses %j → %s", (input, expectedIso) => {
-    expect(parseFootyWireDate(input)?.toISOString()).toBe(expectedIso);
+    expect(parseDate(input)?.toISOString()).toBe(expectedIso);
   });
 
   it.each(["", "2024/03/16", "30 Feb 2024"])("returns null for %j (no defaultYear)", (input) => {
-    expect(parseFootyWireDate(input)).toBeNull();
+    expect(parseDate(input)).toBeNull();
   });
 
   it("parses time during AEDT (UTC+11) correctly", () => {
     // March 13 2025 is during AEDT — 7:30pm AEDT = 08:30 UTC
-    const date = parseFootyWireDate("Thu 13 Mar 7:30pm", 2025);
+    const date = parseDate("Thu 13 Mar 7:30pm", 2025);
     expect(date?.toISOString()).toBe("2025-03-13T08:30:00.000Z");
   });
 
   it("parses time during AEST (UTC+10) correctly", () => {
     // July 13 2025 is during AEST — 7:30pm AEST = 09:30 UTC
-    const date = parseFootyWireDate("13 Jul 7:30pm", 2025);
+    const date = parseDate("13 Jul 7:30pm", 2025);
     expect(date?.toISOString()).toBe("2025-07-13T09:30:00.000Z");
   });
 
   it("parses date-only with defaultYear", () => {
-    expect(parseFootyWireDate("13 Mar", 2025)?.toISOString()).toBe("2025-03-13T00:00:00.000Z");
+    expect(parseDate("13 Mar", 2025)?.toISOString()).toBe("2025-03-13T00:00:00.000Z");
   });
 
   it("parses am time with defaultYear", () => {
     // April 1 2025 is during AEDT — 11:00am AEDT = 00:00 UTC
-    const date = parseFootyWireDate("1 Apr 11:00am", 2025);
+    const date = parseDate("1 Apr 11:00am", 2025);
     expect(date?.toISOString()).toBe("2025-04-01T00:00:00.000Z");
   });
 
   it.each(["13 Mar 7:30pm", "13 Mar"])("returns null for %j without defaultYear", (input) => {
-    expect(parseFootyWireDate(input)).toBeNull();
+    expect(parseDate(input)).toBeNull();
   });
 });
 
-describe("parseAflTablesDate", () => {
+describe("parseDate — AFL Tables inputs", () => {
   it.each([
     ["16-Mar-2024", "2024-03-16T00:00:00.000Z"],
     ["Sat 16-Mar-2024", "2024-03-16T00:00:00.000Z"],
@@ -158,11 +142,11 @@ describe("parseAflTablesDate", () => {
     ["16/Mar/2024", "2024-03-16T00:00:00.000Z"],
     ["8-May-1897", "1897-05-08T00:00:00.000Z"],
   ])("parses %j → %s", (input, expectedIso) => {
-    expect(parseAflTablesDate(input)?.toISOString()).toBe(expectedIso);
+    expect(parseDate(input)?.toISOString()).toBe(expectedIso);
   });
 
   it.each(["", "2024", "16-Xyz-2024"])("returns null for %j", (input) => {
-    expect(parseAflTablesDate(input)).toBeNull();
+    expect(parseDate(input)).toBeNull();
   });
 });
